+++ ccramic/app/inputs/pixel_level_inputs.py
@@ -45,7 +45,7 @@ def render_default_annotation_canvas(input_id: str="annotation_canvas", fullscre
 
     return dash_draggable.GridLayout(id='draggable', children=[canvas]) if draggable else canvas
 
-def wrap_canvas_in_loading_screen_for_large_images(image=None, size_threshold=3000, hovertext=False):
+def wrap_canvas_in_loading_screen_for_large_images(image=None, size_threshold=3000, hovertext=False, enable_zoom=False):
     """
     Wrap the annotation canvas in a dcc.Loading screen if the dimensions of the image are larger than the threshold
     or
@@ -54,10 +54,10 @@ def wrap_canvas_in_loading_screen_for_large_images(image=None, size_threshold=30
     # conditions for wrapping the canvas
     large_image = image is not None and (image.shape[0] > size_threshold or image.shape[1] > size_threshold)
     if large_image or hovertext:
-        return dcc.Loading(render_default_annotation_canvas(),
+        return dcc.Loading(render_default_annotation_canvas(fullscreen_mode=enable_zoom),
                                      type="default", fullscreen=False)
     else:
-        return render_default_annotation_canvas()
+        return render_default_annotation_canvas(fullscreen_mode=enable_zoom)
 
 def add_scale_value_to_figure(figure, image_shape, scale_value=None, font_size=12, x_axis_left=0.05, pixel_ratio=1,
                               invert=False):